Improve separation of concerns
