+++ packages/moxb/src/routing/StateSpaceHandlerImpl.ts
@@ -1,53 +1,12 @@
 import { SubState } from "./StateSpace";
-import { LocationManager } from "./LocationManager";
 import { StateSpaceHandler, StateSpaceHandlerProps } from "./StateSpaceHandler";
-import { UrlArg } from "./UrlArg";
 
 export class StateSpaceHandlerImpl implements StateSpaceHandler {
 
-    private readonly _locationManager: LocationManager;
-    private readonly _substates: SubState[];
-    private readonly _urlArg?: UrlArg<string>;
-    private readonly _rootPath?: string;
+    protected readonly _substates: SubState[];
 
     public constructor(props: StateSpaceHandlerProps) {
-        const { locationManager, substates, rootPath, arg } = props;
-        this._locationManager = locationManager;
-        this._substates = substates;
-        this._urlArg = arg;
-        this._rootPath = rootPath;
-        this.isSubStateActive = this.isSubStateActive.bind(this);
-    }
-
-    public getPathForSubState(state: SubState) {
-        const { root, path } = state;
-        const realRootPath = this._rootPath || this._locationManager.pathSeparator;
-        const result = realRootPath + (root ? '' : path);
-        return result;
-    }
-
-    public isSubStateActive(state: SubState) {
-        const { root, path } = state;
-        if (root || path) {
-            if (this._urlArg) {
-                return this._urlArg.value === path;
-            } else {
-                const toPath = this.getPathForSubState(state);
-                return this._locationManager.isLinkActive(toPath, !!root);
-            }
-        } else {
-            return false;
-        }
-    }
-
-    public getActiveSubStates(): SubState[] {
-        return this._substates
-            .filter(this.isSubStateActive);
-    }
-
-    public getActiveSubStatePaths(): string[] {
-        return this.getActiveSubStates()
-            .map(state => state.path!);
+        this._substates = props.substates;
     }
 
     public findRoot(): SubState {
@@ -70,5 +29,4 @@ export class StateSpaceHandlerImpl implements StateSpaceHandler {
             throw new Error("Can't find subState for path " + path);
         }
     }
-
 }
