+++ packages/moxb/src/routing/index.ts
@@ -34,3 +34,7 @@ export {
 
 export { StateSpaceHandler, StateSpaceHandlerProps } from "./StateSpaceHandler";
 export { StateSpaceHandlerImpl } from "./StateSpaceHandlerImpl";
+
+
+export { StateSpaceAndLocationHandler, StateSpaceAndLocationHandlerProps } from "./StateSpaceAndLocationHandler";
+export { StateSpaceAndLocationHandlerImpl } from "./StateSpaceAndLocationHandlerImpl";
